+++ acceptancetests/jujupy/k8s_provider/__init__.py
@@ -17,75 +17,7 @@
 # (including paths etc.)
 
 
-import logging
-from enum import Enum
-
-logger = logging.getLogger(__name__)
-
-
-class ProviderNotAvailable(AttributeError):
-    """Raised when a provider is requested that isn't registered"""
-    ...
-
-
-class ProviderNotValid(ValueError):
-    """Raised when a provider is not correctly named.
-    always use K8sProviderType to name a new provider.
-    """
-    ...
-
-
-class K8sProviderType(Enum):
-    MICROK8S = 1
-    K8S_CORE = 2
-
-    @classmethod
-    def keys(cls):
-        return list(cls.__members__.keys())
-
-    @classmethod
-    def values(cls):
-        return list(cls.__members__.values())
-
-
-class Factory(object):
-
-    def __init__(self):
-        self._providers = dict()
-
-    def __getitem__(self, name):
-        return self.__getattr__(name)
-
-    def __getattr__(self, name):
-        logger.debug('getting provider %s', name)
-        try:
-            key = K8sProviderType[name]
-            return self._providers[key]
-        except KeyError:
-            raise ProviderNotAvailable("provider {} is not defined".format(key.name))
-
-    @property
-    def providers(self):
-        return self._providers.keys()
-
-    def register(self, provider):
-        key = provider.name
-        if key in self._providers.keys():
-            logger.warn(
-                "provider %s exists. %s will be replaced",
-                key, self._providers[key].name,
-            )
-        self._providers[key] = provider
-
-    def __iter__(self):
-        return iter(self._providers.values())
-
-
-def register_provider(provider):
-    if provider.name not in K8sProviderType.values():
-        raise ProviderNotValid()
-    providers.register(provider)
-    return provider
-
-
-providers = Factory()
+from .base import K8sProviderType  # noqa
+from .kubernetes_core import KubernetesCore  # noqa
+from .microk8s import MicroK8s  # noqa
+from .factory import providers  # noqa
