+++ acceptancetests/jujupy/k8s_provider/microk8s.py
@@ -16,16 +16,17 @@
 # Functionality for handling installed or other juju binaries
 # (including paths etc.)
 
+
 from __future__ import print_function
 
 import logging
 import shutil
 
-from . import (
+from .base import (
     Base,
     K8sProviderType,
-    register_provider,
 )
+from .factory import register_provider
 
 
 logger = logging.getLogger(__name__)
@@ -34,7 +35,7 @@ logger = logging.getLogger(__name__)
 @register_provider
 class MicroK8s(Base):
 
-    name = K8sProviderType.MicroK8s
+    name = K8sProviderType.MICROK8S
 
     def __init__(self, client, timeout=1800):
         super().__init__(client, timeout)
